test(conway): Addressed all the comments of the PR and moved the tests to block_test.go
